pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Stefan Scheffler <stefanscheffler82@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11.9,<3.14"
pillow = "^11.1.0"
pillow-heif = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
